// console/src/generated/UsageRecordQueryDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type UsageRecordQueryDto = { from: number, to: number, user_key_id: number | null, user_id: number | null, provider_id: number | null, credential_id: number | null, model: string | null, request_id: string | null, operation: string | null, usage_source: string | null, ended: string | null, page: number | null, page_size: number | null, include_total?: boolean, };